+++ src/deployer/adapters/TruffleAdapter.ts
@@ -43,7 +43,6 @@ export class TruffleAdapter extends Adapter {
     );
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public async toInstance<I>(instance: TruffleFactory<I>, address: string, _: OverridesAndName): Promise<I> {
     const contract = this._hre.artifacts.require(instance.contractName!);