+++ src/tools/reporters/PublicReporter.ts
@@ -1,3 +1,7 @@
+import { format } from "prettier";
+
+import { Reporter } from "./Reporter";
+
 /* eslint-disable no-console */
 import { networkManager } from "../network/NetworkManager";
 
@@ -24,5 +28,30 @@ export class PublicReporter {
     console.log();
     console.table(table);
     console.log();
+  }
+
+  public static async reportContractsMD(...contracts: [string, string][]): Promise<void> {
+    const explorer = await Reporter!.getExplorerUrl();
+    const normalizedExplorer = explorer.endsWith("/") ? explorer : `${explorer}/address/`;
+
+    let markdownContent;
+
+    const headers = [
+      "| Contract      | Address                                    |",
+      "|---------------|--------------------------------------------|",
+    ];
+    const rows = contracts.map(([contract, address]) => {
+      const shortenedAddress = `${address.slice(0, 6)}...${address.slice(-4)}`; // Shorten address
+      return `| ${contract} | [${shortenedAddress}](${normalizedExplorer}${address}) |`;
+    });
+
+    markdownContent = [...headers, ...rows].join("\n");
+    console.log(
+      await format(markdownContent, {
+        parser: "markdown",
+        printWidth: 80,
+        proseWrap: "always",
+      }),
+    );
   }
 }