+++ src/tools/runners/TransactionRunner.ts
@@ -42,7 +42,7 @@ class BaseTransactionRunner {
 
     await Reporter!.reportTransactionReceipt(receipt);
 
-    this.totalCost += BigInt(receipt.fee) + BigInt(tx.value) ?? 0n;
+    this.totalCost += BigInt(receipt.fee ?? 0n) + BigInt(tx.value ?? 0n);
     this.totalTransactions++;
   }