+++ src/utils/common.ts
@@ -1,13 +1,13 @@
-import { join } from "path";
-import { realpathSync, existsSync } from "fs";
+import { join, sep } from "path";
+
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
 import { UNKNOWN_CONTRACT_NAME } from "../constants";
 
-export function resolvePathToFile(path: string, file: string = ""): string {
-  if (!existsSync(path)) {
-    path = "./";
-  }
+export function resolvePathToFile(hre: HardhatRuntimeEnvironment, path: string, file: string = ""): string {
+  const pathToMigration = join(hre.config.paths.root, path, file);
 
-  return join(realpathSync(path), file);
+  return pathToMigration.endsWith(sep) ? pathToMigration.slice(0, -1) : pathToMigration;
 }
 
 export function getInstanceNameFromClass(instance: any): string {