+++ test/unit/deploy-files/1_mock.migration.ts
@@ -1,3 +1,0 @@
-import { Deployer } from "../../../src/deployer/Deployer";
-
-export = async (deployer: Deployer) => {};